backend/requirements.txt
fastapi==0.115.8
uvicorn[standard]==0.34.0
SQLAlchemy==2.0.37
psycopg2-binary==2.9.10
python-jose[cryptography]==3.3.0
passlib==1.7.4
pydantic-settings==2.7.1
email-validator==2.2.0
openai==1.59.7
qdrant-client==1.12.1
python-multipart==0.0.20
